pyproject.toml
[project]
name = "promptdev"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "psycopg2-binary",
    "prometheus-client",
    "pytest",
    "testcontainers[postgres]>=4.13.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
